requirements.txt
mysql-connector-python==8.3.0
boto3==1.34.69
python-dotenv==1.0.1
twilio==9.0.5
jsonschema==4.21.1
